+++ src/MyNeuralNetwork.Domain.Tests/Entities/Support/PropertiesTester.cs
@@ -14,7 +14,7 @@ namespace MyNeuralNetwork.Domain.Tests.Entities.Support
             Assert.That(neuron.Value.Value, Is.EqualTo(expectedValue));
         }
 
-        public static void TestBias(Neuron neuron, float expectBias)
+        public static void TestBias(Neuron neuron, double expectBias)
         {
             TestContext.Write("Testing if bias is ");
             TestContext.WriteLine(expectBias);
@@ -22,7 +22,7 @@ namespace MyNeuralNetwork.Domain.Tests.Entities.Support
             Assert.That(neuron.Bias, Is.EqualTo(expectBias));
         }
 
-        public static void TestGamma(Neuron neuron, float expectedGamma)
+        public static void TestGamma(Neuron neuron, double expectedGamma)
         {
             TestContext.Write("Testing if gamma is ");
             TestContext.WriteLine(expectedGamma);
@@ -30,7 +30,7 @@ namespace MyNeuralNetwork.Domain.Tests.Entities.Support
             Assert.That(neuron.Gamma, Is.EqualTo(expectedGamma));
         }
 
-        public static void TestWeight(Neuron neuron, Neuron nextNeuron, float expectValue)
+        public static void TestWeight(Neuron neuron, Neuron nextNeuron, double expectValue)
         {
             TestContext.Write("Testing if weight is ");
             TestContext.WriteLine(expectValue);
